ajustes finos da tela de habits

+++ src/components/Habits.js
@@ -5,6 +5,7 @@ import CreateHabit from "./CreateHabit"
 
 import Style from "./styledComponents/Style"
 import styled from "styled-components"
+import { ThemeProvider } from "styled-components"
 import trashIcon from "../assets/imgs/Vectortrash.png"
 
 import UserContext from "./Context/UserContext"
@@ -15,7 +16,7 @@ import { useEffect, useState, useContext } from "react"
 
 
 export default function Habits() {
-    const {habits, setHabits, addHabit, setAddHabbit} = useContext(HabitsContext)
+    const { habits, setHabits, addHabit, setAddHabbit } = useContext(HabitsContext)
     const { user } = useContext(UserContext)
 
     const token = localStorage.getItem("token")
@@ -27,47 +28,64 @@ export default function Habits() {
             },
         };
         const URL = "https://mock-api.bootcamp.respondeai.com.br/api/v2/trackit/habits"
-        axios.get(URL, config).then(response => { setHabits(response.data)
+        axios.get(URL, config).then(response => {
+            setHabits(response.data)
         }
         ).catch(err => { console.log("deu ruim no get dos hábitos", err.response.data) })
 
     }, [addHabit])
+
     function renderHabits() {
+        const weekDays = [
+            { letter: 'D', number: 7 },
+            { letter: 'S', number: 1 },
+            { letter: 'T', number: 2 },
+            { letter: 'Q', number: 3 },
+            { letter: 'Q', number: 4 },
+            { letter: 'S', number: 5 },
+            { letter: 'S', number: 6 }
+        ];
+        console.log(habits)
         return (
-            habits.map((habit,index) => {
+            habits.map((habit, index) => {
                 return <div className="card" key={index} >
                     <div className="title" >
-                    <p>{habit.name}</p>
-                    <img src={trashIcon} onClick={()=>deleteHabit(habit.id,habit.name)} />
+                        <p>{habit.name}</p>
+                        <img src={trashIcon} onClick={() => deleteHabit(habit.id, habit.name)} />
                     </div>
                     <Weekdays>
-                        {habit.days.map((day,index) => {
-                            if (day === 7) { return <div className="weekday"><p>D</p></div> }
-                            else if (day === 1) { return <div className="weekday"><p>S</p></div> }
-                            else if (day === 2) { return <div className="weekday"><p>T</p></div> }
-                            else if (day === 3 || day === 4) { return <div className="weekday"><p>Q</p></div> }
-                            else if (day === 5 || day === 6) { return <div className="weekday"><p>S</p></div> }
+                        {weekDays.map((day) => {
+                            return <div className="weekday">
+                                <ThemeProvider theme={habit.days.includes(day.number) ? invertedColor : color} >
+                                    <Span>{day.letter}</Span>
+                                 </ThemeProvider>
+                            </div>
+
                         })}
                     </Weekdays>
-
                 </div>
             })
         )
     }
 
-    function deleteHabit(id,hab){
+    function deleteHabit(id, hab) {
         const token = localStorage.getItem("token")
-        const  URL_DELETE = `https://mock-api.bootcamp.respondeai.com.br/api/v2/trackit/habits/${id}`
+
+        const URL_DELETE = `https://mock-api.bootcamp.respondeai.com.br/api/v2/trackit/habits/${id}`
         const config = {
             headers: {
                 Authorization: `Bearer ${token}`,
             },
         };
-        axios.delete(URL_DELETE,config).then(console.log("deletei o habito ", hab)).catch(err => {console.log(err.response)})
 
-            let newHabits = habits.filter((h)=>h.id!==id)
-            setHabits(newHabits) 
-    }
+     if(window.confirm("Você realmente deseja apagar esse hábito?")){
+         axios.delete(URL_DELETE, config).then(console.log("deletei o habito ", hab)).catch(err => { console.log(err.response) })
+ 
+         let newHabits = habits.filter((h) => h.id !== id)
+         setHabits(newHabits)
+ 
+     }
+     }
 
     return (
         <>
@@ -77,7 +95,7 @@ export default function Habits() {
                     <p>Meus Hábitos</p>
                     <button onClick={() => { setAddHabbit(!addHabit); console.log("addhabit", addHabit) }} >+</button>
                 </div>
-                {addHabit ? <CreateHabit/> : <></>}
+                {addHabit ? <CreateHabit /> : <></>}
                 {habits.length > 0 ? renderHabits() :
                     <span>Você não tem nenhum hábito cadastrado ainda. Adicione um hábito para começar a trackear!</span>
                 }
@@ -88,6 +106,25 @@ export default function Habits() {
 }
 
 
+const Span = styled.div`
+color: ${props => props.theme.dfColor};
+background-color: ${props => props.theme.dfBack};
+display: flex;
+justify-content: center;
+align-items: center;
+width: 29px;
+ height: 28px;
+    
+`
+const color = {
+    dfColor: '#D4D4D4',
+    dfBack: '#FFFFFF'
+};
+
+const invertedColor = {
+    dfColor: '#FFFFFF',
+    dfBack: '#D4D4D4'
+};
 
 const Weekdays = styled.div` 
 display: flex;
@@ -96,6 +133,7 @@ display: flex;
     width: 30px;
     height: 30px;
     color: #dbdbdb;
+    background-color: #dbdbdb;
     border: solid 1px #d4d4d4;
     border-radius: 3px;
     margin: 10px 3px;
